Added auth header to the secure workout routes

+++ Client/src/components/WorkoutDetails.jsx
@@ -1,14 +1,20 @@
 import axios from "axios";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { formatDistanceToNow } from "date-fns";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if (!user) {
+            return;
+        }
         try {
             const response = await axios.delete(
-                "http://localhost:3000/api/workouts/" + workout._id
+                "http://localhost:3000/api/workouts/" + workout._id,
+                { headers: { Authorization: `Bearer ${user.token} ` } }
             );
 
             if (response) {
